requirements.txt
###### Requirements ######
flask
flask_socketio
RPi.GPIO
GPIO
numpy
godirect
scipy
markdown
gunicorn 
eventlet